esp32 wifi refactor to push network event consumption out of the coprocessor
