+++ source/implementations/f7/Meadow.F7/Devices/F7MicroBase.cs
@@ -89,17 +89,7 @@ public abstract partial class F7MicroBase : IF7MeadowDevice
         var reservedPins = PlatformOS.ReservedPins?.ToUpper().Split(';', StringSplitOptions.RemoveEmptyEntries) ?? null;
         IoController.Initialize(reservedPins);
 
-        InitializeCellModem();
         InitCoprocessor();
-    }
-
-    private void InitializeCellModem()
-    {
-        if (PlatformOS.SelectedNetwork == IPlatformOS.NetworkConnectionType.Cell)
-        {
-            Resolver.Log.Info($"Device is configured to use Cell for the network interface");
-            networkAdapters.Add(new F7CellNetworkAdapter());
-        }
     }
 
     /// <summary>
@@ -123,21 +113,27 @@ public abstract partial class F7MicroBase : IF7MeadowDevice
                     if (PlatformOS.SelectedNetwork == IPlatformOS.NetworkConnectionType.WiFi)
                     {
                         Resolver.Log.Info($"Device is configured to use WiFi for the network interface");
-                        networkAdapters.Add(esp32);
+                        var wifiAdapter = new Esp32WiFiAdapter(esp32);
+                        networkAdapters.Add(wifiAdapter);
 
-                        if (esp32.AutoConnect)
+                        if (wifiAdapter.AutoConnect)
                         {
-                            Resolver.Log.Debug($"Device configured to auto-connect to SSID '{esp32.DefaultSsid}'");
+                            Resolver.Log.Debug($"Device configured to auto-connect to SSID '{wifiAdapter.DefaultSsid}'");
 
-                            if (string.IsNullOrEmpty(esp32.DefaultSsid))
+                            if (string.IsNullOrEmpty(wifiAdapter.DefaultSsid))
                             {
                                 Resolver.Log.Warn($"Device configured to auto-connect to WiFi, but no Default SSID was provided.  Deploy a wifi.config.yaml file.");
                             }
                             else
                             {
-                                esp32.ConnectToDefaultAccessPoint();
+                                wifiAdapter.ConnectToDefaultAccessPoint();
                             }
                         }
+                    }
+                    else if (PlatformOS.SelectedNetwork == IPlatformOS.NetworkConnectionType.Cell)
+                    {
+                        Resolver.Log.Info($"Device is configured to use Cell for the network interface");
+                        networkAdapters.Add(new F7CellNetworkAdapter(esp32));
                     }
                 }
                 catch (Exception e)
